// ui/staff-ui/src/features/change-request/types/change-request.ts
import { UploadedDocument } from "@/features/shared/types";

export interface ChangeDocument {
    doc_id: string;
    doc_name: string;
    doc_url?: string;
}

export interface ChangeRequest {
    change_request_id: string;
    record_name?: string | null;
    register_id: string;
    register_mnemonic?: string | null;
    tab_id: string;
    tab_label?: string | null;
    internal_record_id: string;
    section_id: string;
    section_mnemonic?: string;
    section_register_id: string;

    source_partner_id: string;
    created_by: string;
    created_at: string;

    approval_status: 'PENDING' | 'APPROVED' | 'REJECTED';
    approved_by: string | null;
    approved_at: string | null;

    no_of_verifications_required: number;
    no_of_verifications_done: number;
    awe_request_id?: string | null;
    awe_request_status_summary?: string | null;
    is_list: boolean;
    change_payload: any;
    current_register_data: any;
    documents: UploadedDocument[] | null;
    is_core_section: boolean;
    is_primary_section: boolean;
}